v2_0/requirements.txt
absl-py==1.2.0
addict==2.4.0
astor==0.8.1
attrdict==2.0.1
Babel==2.10.3
backcall==0.2.0
bce-python-sdk==0.8.74
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cssselect==1.1.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.2
decorator==5.1.1
dill==0.3.5.1
entrypoints==0.4
et-xmlfile==1.1.0
Flask==2.2.2
Flask-Babel==2.0.0
fonttools==4.37.2
future==0.18.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.49.0
idna==3.4
imageio==2.21.3
imgaug==0.4.0
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==7.34.0
itsdangerous==2.1.2
jarowinkler==1.2.1
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.11.1
kiwisolver==1.4.4
lanms-neo==1.0.2
lmdb==1.3.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
mmcv-full==1.6.1
mmdet==2.25.2
mmocr==0.6.1
model-index==0.1.11
multiprocess==0.70.13
nest-asyncio==1.5.5
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openmim==0.3.1
openpyxl==3.0.10
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==21.3
paddle-bfloat==0.1.7
paddleocr==2.6.0.1
paddlepaddle==2.3.2
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
premailer==3.10.0
prompt-toolkit==3.0.30
protobuf==3.19.5
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post3
pycocotools==2.0.5
pycryptodome==3.15.0
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
pywin32==304
PyYAML==6.0
pyzmq==23.2.0
rapidfuzz==2.9.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
seaborn==0.12.0
Shapely==1.8.4
six==1.16.0
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
tifffile==2021.11.2
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
visualdl==2.4.1
wcwidth==0.2.5
Werkzeug==2.2.2
wincertstore==0.2
yapf==0.32.0
zipp==3.8.1
